+++ docs/source/whats_new.rst
@@ -32,7 +32,7 @@ capabilities, including:
     :width: 300pt
 
 * The SRTM IO got some new features: :func:`cartopy.io.srtm.add_shading` adds
-  shades on SRTM elevation maps. :fun:c`cartopy.io.srtm.fill_gaps` might be
+  shades on SRTM elevation maps. :func:`cartopy.io.srtm.fill_gaps` might be
   needed if SRTM has gaps, but requires `gdal` to be installed.
 
 What's new in cartopy 0.9